Refactored AddNewBooksForm
