+++ src/pages/AddNewBooksForm/BookForm/index.ts
@@ -1,0 +1,1 @@
+export { default as BookForm } from './BookForm'
