+++ src/pages/AddNewBooksForm/CustomComponents/index.ts
@@ -1,0 +1,3 @@
+export { default as CustomFormLabel } from './CustomFormLabel'
+export { default as CustomTextField } from './CustomTextField'
+export { default as RemoveBookButton } from './RemoveBookButton'
